Try to calculate current with R_SENSE of 18mOhms but still no luck.

+++ src/drivers/pac195x/pac195x.h
@@ -67,6 +67,6 @@ int pac195x_refresh_v(SensorLocation const *loc);
 int pac195x_refresh_g(SensorLocation const *loc);
 
 uint32_t pac195x_calc_bus_voltage(uint8_t fsr, uint16_t vbus, bool bipolar);
-uint32_t pac195x_calc_bus_current(uint16_t rsense, uint16_t vsense, bool bipolar);
+uint32_t pac195x_calc_bus_current(uint32_t rsense, uint16_t vsense, bool bipolar);
 
 #endif // _PAC195X_H_
